+++ sdk/src/wallet/core/operations/stronghold.rs
@@ -17,8 +17,10 @@ impl Wallet {
 
         if let SecretManager::Stronghold(stronghold) = &mut *self.secret_manager.write().await {
             stronghold.set_password(password).await?;
+            Ok(())
+        } else {
+            Err(crate::client::Error::SecretManagerMismatch.into())
         }
-        Ok(())
     }
 
     /// Change the Stronghold password to another one and also re-encrypt the values in the loaded snapshot with it.
@@ -33,44 +35,50 @@ impl Wallet {
         if let SecretManager::Stronghold(stronghold) = &mut *self.secret_manager.write().await {
             stronghold.set_password(current_password).await?;
             stronghold.change_password(new_password).await?;
+            Ok(())
+        } else {
+            Err(crate::client::Error::SecretManagerMismatch.into())
         }
-        Ok(())
     }
 
     /// Sets the Stronghold password clear interval
     pub async fn set_stronghold_password_clear_interval(&self, timeout: Option<Duration>) -> crate::wallet::Result<()> {
         if let SecretManager::Stronghold(stronghold) = &mut *self.secret_manager.write().await {
             stronghold.set_timeout(timeout).await;
+            Ok(())
+        } else {
+            Err(crate::client::Error::SecretManagerMismatch.into())
         }
-        Ok(())
     }
 
     /// Stores a mnemonic into the Stronghold vault
     pub async fn store_mnemonic(&self, mnemonic: Mnemonic) -> crate::wallet::Result<()> {
         if let SecretManager::Stronghold(stronghold) = &mut *self.secret_manager.write().await {
             stronghold.store_mnemonic(mnemonic).await?;
+            Ok(())
+        } else {
+            Err(crate::client::Error::SecretManagerMismatch.into())
         }
-        Ok(())
     }
 
     /// Clears the Stronghold password from memory.
     pub async fn clear_stronghold_password(&self) -> crate::wallet::Result<()> {
         log::debug!("[clear_stronghold_password]");
-        let mut secret_manager = self.secret_manager.write().await;
-        match &mut *secret_manager {
-            SecretManager::Stronghold(stronghold) => stronghold.clear_key().await,
-            _ => return Err(crate::client::Error::SecretManagerMismatch.into()),
+        if let SecretManager::Stronghold(stronghold) = &mut *self.secret_manager.write().await {
+            stronghold.clear_key().await;
+            Ok(())
+        } else {
+            Err(crate::client::Error::SecretManagerMismatch.into())
         }
-        Ok(())
     }
 
     /// Checks if the Stronghold password is available.
     pub async fn is_stronghold_password_available(&self) -> crate::wallet::Result<bool> {
         log::debug!("[is_stronghold_password_available]");
-        let mut secret_manager = self.secret_manager.write().await;
-        match &mut *secret_manager {
-            SecretManager::Stronghold(stronghold) => Ok(stronghold.is_key_available().await),
-            _ => Err(crate::client::Error::SecretManagerMismatch.into()),
+        if let SecretManager::Stronghold(stronghold) = &*self.secret_manager.write().await {
+            Ok(stronghold.is_key_available().await)
+        } else {
+            Err(crate::client::Error::SecretManagerMismatch.into())
         }
     }
 }