+++ sdk/CHANGELOG.md
@@ -24,10 +24,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Wallet::get_or_create_account` convenience method;
+- `Output::kind_str()` method;
 
 ### Changed
 
 - `StrongholdAdapterBuilder` updated to be slightly more ergonomic;
+- `Wallet::{set_stronghold_password, change_stronghold_password, set_stronghold_password_clear_interval, store_mnemonic}` return an `Err` instead of `Ok` in case of a non-stronghold secret manager; 
 
 ## 1.0.1 - 2023-07-25